Move get_graphql_schema_from_orientdb_schema_data to a more suitable location
